001-AppHack/001-AppUpload/requirements.txt
streamlit==1.23.0
azure-storage-blob==12.10.0
azure-servicebus==7.6.0
